+++ apps/web/src/hooks/hotkeys/use-global-hotkeys.ts
